Many fixes and improvements Added proper camera orientation to calibration normalisation Made async detection interruptible if target was detected Added grace period before regarding target as lost to increase resilience to lost frames Greatly improved brute force detection: - Better candidate valuation at distance where points < projections - Switched to slow point matching, using fast one was a bug Added origin cross visualisation Added camera pose to VRPN output as reference Fixed VRPN coordinate system handedness Fixed LTO build Fixed comment typos Fixed many crashes

+++ shared/util/util.hpp
@@ -105,7 +105,7 @@ static inline INT shortDiff(UINT a, UINT b, INT bias, INT overflow)
 {
 	static_assert(std::numeric_limits<UINT>::lowest() == 0);
 	assert(a < overflow && b < overflow);
-	INT passed = b - a;
+	INT passed = ((INT)b) - a;
 	if (passed < -bias)
 		passed += overflow;
 	else if (passed > overflow-bias)
